Step 7: Extract git history scanning to scanner_history.py
Moves scan_git_history, should_skip, try_decode, and helpers _build_blob_map
and _read_blob into scanner_history.py. scanner.py becomes a thin orchestrator.
